Front End of Admin Company View
Added two custom components to Admin Company View:
- Owner Individual
- Employee Individual

also added a data link between the parent and child.

Issue: We need to be able to generate multiples of the same component. meaning that the datalink needs to change in order for each employee to get their own name. currently all the components is linked to the same variable and will have the same name, Cornel does not yet know how to fix this issue